pyproject.toml
[project]
name = "mycloud"
version = "0.1.0"
description = ""
authors = [
    {name = "kihoon",email = "nike6736@naver.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.8,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
